pyproject.toml
[tool.poetry]
name = "data-engineering"
version = "0.1.0"
description = ""
authors = ["Makhonin Alexey <makhonin.a.ru@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.13"
pandas = "^2.2.2"
tqdm = "^4.66.2"
types-tqdm = "^4.66.0.20240417"
pandas-stubs = "^2.2.1.240316"
apache-airflow = {extras = ["cncf-kubernetes"], version = "^2.9.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
